+++ src/main/java/com/nhnacademy/bookstoreuserapi/guest/controller/GuestController.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.bookstoreuserapi.guest.controller;
 
-import com.nhnacademy.bookstoreuserapi.guest.domain.GuestCreateRequest;
-import com.nhnacademy.bookstoreuserapi.guest.domain.GuestUpdateRequest;
-import com.nhnacademy.bookstoreuserapi.guest.domain.ResponseGuest;
+import com.nhnacademy.bookstoreuserapi.guest.dto.request.GuestCreateRequest;
+import com.nhnacademy.bookstoreuserapi.guest.dto.request.GuestUpdateRequest;
+import com.nhnacademy.bookstoreuserapi.guest.dto.response.ResponseGuest;
 import com.nhnacademy.bookstoreuserapi.common.exception.ValidationFailedException;
 import com.nhnacademy.bookstoreuserapi.guest.service.GuestService;
 import jakarta.validation.Valid;
